+++ R/cohort_creation.R
@@ -90,23 +90,24 @@
 #'
 #' @export
 cohort_creation <- function(
-    cohort,
-    labels,
-    exclusion_flag = NULL,
-    show_prct = TRUE,
-    group_var = NULL,
-    ...) {
+  cohort,
+  labels,
+  exclusion_flag = NULL,
+  show_prct = TRUE,
+  group_var = NULL,
+  ...
+) {
   ## if no exclusion flags provided, interpret all steps as "inclusions"
   if (is.null(exclusion_flag)) {
     exclusion_flag <- c(rep(FALSE, length(cohort)))
   }
 
   ## check user input
-  Rgemini:::check_input(cohort, "list")
-  Rgemini:::check_input(labels, "character")
-  Rgemini:::check_input(list(exclusion_flag, show_prct), "logical")
+  check_input(cohort, "list")
+  check_input(labels, "character")
+  check_input(list(exclusion_flag, show_prct), "logical")
   if (!is.null(group_var)) {
-    Rgemini:::check_input(group_var, "character")
+    check_input(group_var, "character")
     ## check if group columns exist in all cohort list items
     lapply(cohort, function(cohort) {
       check_input(cohort, "data.table", colnames = group_var)